requirements.txt
gunicorn
openai
pinecone-client
python-dotenv
streamlit
streamlit-chat